+++ CHANGELOG.md
@@ -6,3 +6,5 @@ This is only for new-version changes:
 
 * added the `undo` command
 * added the `status` command
+* added the `configure` command
+* added the `show` command